{#40} url 분리
