+++ http-test/user-api.http
@@ -18,6 +18,6 @@ Authorization: Bearer {{jwtToken}}
 
 
 ### 유저 통계 조회
-GET http://localhost:8080/user/v1/user/statistics
+GET http://localhost:8080/admin/v1/user/statistics
 Content-Type: application/json
 Authorization: Bearer {{jwtToken}}
